refactor(api): create controllers to be called directly from server components
