--- src/app/api/review/[id]/types.ts
@@ -1,4 +1,0 @@
-import { z } from "zod";
-import { reviewSchema } from "./schema";
-
-export type Review = z.infer<typeof reviewSchema>;
