+++ src/app/api/review/route.ts
@@ -1,17 +1,8 @@
-import prisma from "@/lib/prisma";
-import { errorHandler } from "@/middleware/api/errorHandler";
 import { NextResponse } from "next/server";
-import { reviewListSchema } from "./schema";
-import { ReviewList } from "./types";
+import { ReviewController } from "@/services/controllers/review";
 
 export async function GET() {
-  try {
-    const reviews = await prisma.review.findMany({ include: { user: true } });
+  const reviews = await ReviewController.listAll();
 
-    const filteredReviews: ReviewList = reviewListSchema.parse(reviews);
-
-    return NextResponse.json({ data: filteredReviews });
-  } catch (error) {
-    return errorHandler(error);
-  }
+  return NextResponse.json(reviews);
 }
