+++ src/services/controllers/review/schema.ts
@@ -1,4 +1,5 @@
 import { z } from "zod";
+import { getYear } from "@/utils/getYear";
 
 export const reviewListSchema = z.array(
   z.object({
@@ -40,3 +41,43 @@ export const createReviewData = z.object({
     })
     .min(1, "Field required"),
 });
+
+export const reviewSchema = z
+  .object({
+    id: z.string(),
+    title: z.string(),
+    description: z.string(),
+    rating: z.number(),
+    user: z.object({
+      name: z.string(),
+      avatarUrl: z.string(),
+    }),
+    movie: z.object({
+      poster: z.object({
+        src: z.string().nullable(),
+        base64: z.string(),
+      }),
+      title: z.string(),
+      releaseDate: z.string(),
+      genres: z.array(
+        z.object({
+          id: z.number(),
+          name: z.string(),
+        })
+      ),
+    }),
+  })
+  .transform((review) => {
+    const {
+      movie: { releaseDate, ...movieRest },
+      ...rest
+    } = review;
+
+    return {
+      ...rest,
+      movie: {
+        ...movieRest,
+        releaseYear: getYear(releaseDate),
+      },
+    };
+  });
